requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
blinker==1.4              # via -r requirements.in
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask, python-dotenv
flask-oauthlib==0.9.4     # via -r requirements.in
flask-wtf==0.14.3         # via -r requirements.in
flask==0.12.2             # via -r requirements.in, flask-oauthlib, flask-wtf
gunicorn==19.7.1          # via -r requirements.in
idna==2.6                 # via requests
itsdangerous==0.24        # via flask, flask-wtf
jinja2==2.9.6             # via flask
markupsafe==1.0           # via jinja2, wtforms
oauthlib==2.0.6           # via flask-oauthlib, requests-oauthlib
python-dotenv==0.7.1      # via -r requirements.in
raven==6.3.0              # via -r requirements.in
requests-oauthlib==0.8.0  # via flask-oauthlib
requests==2.18.4          # via requests-oauthlib
urllib3==1.22             # via requests
werkzeug==0.12.2          # via flask
whitenoise==5.1.0         # via -r requirements.in
wtforms==2.3.3            # via flask-wtf
